+++ test/recording_studio_notifications_push_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class RecordingStudioNotificationsPushTest < Minitest::Test
   def test_version_matches_release
-    assert_equal "0.1.0", ::RecordingStudioNotificationsPush::VERSION
+    assert_equal "0.1.1", ::RecordingStudioNotificationsPush::VERSION
   end
 
   def test_engine_exists
@@ -76,5 +76,34 @@ class RecordingStudioNotificationsPushTest < Minitest::Test
     )
     assert_includes push_devices_js, "serviceWorkerRegistration"
     assert_includes push_devices_js, "resolveServiceWorkerRegistration"
+    assert_includes push_devices_js, "not mounted"
+    assert_includes push_devices_js, "data-push-enable"
+    assert_includes push_devices_js, "data-push-disable"
+    assert_includes push_devices_js, "detectCurrentBrowser"
+    assert_includes push_devices_js, "showDisable"
+
+    sw_partial = File.read(
+      File.expand_path("../app/views/recording_studio_notifications_push/_service_worker_push.js.erb", __dir__)
+    )
+    assert_includes sw_partial, "showNotification"
+    assert_includes sw_partial, "/icon.png"
+    refute_includes sw_partial, "rsnp:push"
+
+    devices_show = File.read(
+      File.expand_path("../app/views/recording_studio_notifications_push/devices/show.html.erb", __dir__)
+    )
+    assert_includes devices_show, "push_enable: true"
+    assert_includes devices_show, "push_disable: true"
+    assert_includes devices_show, "Disable on this browser"
+  end
+
+  def test_dummy_pwa_head_resolves_service_worker_via_main_app
+    head = File.read(
+      File.expand_path("dummy/app/views/recording_studio/_default_layout_head.html.erb", __dir__)
+    )
+
+    assert_includes head, "main_app.pwa_service_worker_path"
+    assert_includes head, "main_app.pwa_manifest_path"
+    assert_includes head, "RecordingStudioPwa.serviceWorkerReady"
   end
 end